Tested and debugged the FileHandlers for the Pascal dataset,
I think we may need to discuss how to handle the databasets, this Pascal
is quite a mess. We'll talk about it.

I also created a file for mapping (all) the classes (present in a given
scribble directory) to integers, but as i just mentioned, we need to
discuss which one to use before doing it.